% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adj_lw.R
\name{adj_lw_sp}
\alias{adj_lw_sp}
\title{adj_lw_sp}
\usage{
adj_lw_sp(object, neighbour_type = "queen", lw_style = "B", combine = F)
}
\arguments{
\item{object}{an sp object \linkS4class{sp}.}

\item{neighbour_type}{default queen, may also be rook. \code{\link[spdep]{cell2nb}}}

\item{lw_style}{style can take values “W”, “B”, “C”, “U”, “minmax” and “S”. \code{\link[spdep]{nb2listw}}}

\item{combine}{default FALSE, if TRUE all individuals' neighbours list will be consolidated into one large list.}
}
\value{
if combine=F, a value will be assigned to the list_neighbour slot corresponding to the sp object.
if combine=T, a value will be assigned to the combine_listw slot corresponding to the sp object.
}
\description{
Create neighbours list with spatial weights for the chosen coding scheme for the given sp object.
}
