% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adj_matrix.R
\name{adj_matrix_sp}
\alias{adj_matrix_sp}
\title{adj_matrix_sp}
\usage{
adj_matrix_sp(object, neighbour_type = "queen", code_style = "W")
}
\arguments{
\item{object}{an sp object \linkS4class{sp}.}

\item{neighbour_type}{default queen, may also be rook. \code{\link[spdep]{cell2nb}}}

\item{code_style}{style can take values W, B, C, and S. \code{\link[spdep]{nb2mat}}}
}
\value{
a value will be assigned to the adj_matrix_list slot corresponding to the sp object.
}
\description{
Create neighbours list with spatial weights for the chosen coding scheme for the given sp object.
}
