% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap}
\alias{bootstrap}
\alias{bootstrap.data.frame}
\title{bootstrap}
\usage{
bootstrap(x, ...)

\method{bootstrap}{data.frame}(x, group, dif = 1.5, R = 500, ...)
}
\arguments{
\item{x}{A protein expression matrix, where the row names are protein
or gene IDs and the column names are sample names.}

\item{...}{Arguments passed to other methods.}

\item{group}{The grouping information of samples, where each group contains samples from different individuals.}

\item{dif}{The range of the confidence interval.}

\item{R}{The number of bootstrap replicates.}
}
\value{
A protein expression matrix after removing some proteins with large fluctuations..
}
\description{
Use the bootstrap method to calculate the confidence intervals of each protein
in different groups, and filter out proteins with larger thresholds to control
the impact of individual differences on data analysis.
}
\details{
Since proteomics data is generally log-transformed, the range of the confidence interval—i.e.,
the difference between the two ends of the confidence interval—can be interpreted as the fold
change between these two ends.
}
