% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap_sp}
\alias{bootstrap_sp}
\title{bootstrap_sp}
\usage{
bootstrap_sp(object, group, dif = 1.5, R = 500)
}
\arguments{
\item{object}{an sp object \linkS4class{sp}.}

\item{group}{Group information of samples, which is a string corresponding
to one element in the column name of the sp object slot "sample_features".
The attribute of this column should be a character vector or a factor.}

\item{dif}{The range of the confidence interval.}

\item{R}{The number of bootstrap replicates.}
}
\value{
The sp object with the slot RID_bootstrap.
}
\description{
Use the bootstrap method to calculate the confidence intervals of each protein
in different groups, and filter out proteins with larger thresholds of sp object to control
the impact of individual differences on data analysis.
}
\seealso{
\code{\link{bootstrap}}
}
