% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{cluster_read_sp}
\alias{cluster_read_sp}
\title{cluster_read_sp}
\usage{
cluster_read_sp(object, name = "clean", n_cluster, col_name)
}
\arguments{
\item{object}{\linkS4class{sp}.}

\item{name}{keep consistency with the setting of the name parameter in the cluster_sp function.}

\item{n_cluster}{integer value. number of clusters.}

\item{col_name}{the name of the newly generated column.}
}
\description{
Select the results from the clustering files and add a new column to the sample feature data frame of the sp object.
}
