% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{cluster_sp}
\alias{cluster_sp}
\title{cluster_sp}
\usage{
cluster_sp(
  object,
  name = "clean",
  dimn,
  maxK = 10,
  seed = 1234,
  reps = 5000,
  pItem = 0.8,
  pFeature = 1,
  clusterAlg = "km"
)
}
\arguments{
\item{object}{an sp object \linkS4class{sp}.}

\item{name}{There are three options in total:
\itemize{ \item clean :
  The pca results from the clean_data
  \item RID : The pca results of the data with individual differences removed through the bootstrap method (RID_bootstrap).
  \item harmony : The pca results of the data with individual differences removed through the harmony method (corrective_PCA).}}

\item{dimn}{Select the expression matrix of the first n dimensions from the
pca results for ConsensusClusterPlus function.}

\item{maxK}{integer value. maximum cluster number to evaluate.}

\item{seed}{optional numerical value. sets random seed for reproducible results.}

\item{reps}{integer value. number of subsamples.}

\item{pItem}{numerical value. proportion of items to sample.}

\item{pFeature}{numerical value. proportion of features to sample.}

\item{clusterAlg}{character value. cluster algorithm. 'hc' hierarchical (hclust), 'pam' for paritioning
around medoids, 'km' for k-means upon data matrix, or a function that returns a clustering. See example
and vignette for more details.}
}
\value{
A folder containing clustering results.
}
\description{
Execute the ConsensusClusterPlus function for the given sp object.
}
\seealso{
\code{\link[ConsensusClusterPlus]{ConsensusClusterPlus}}
}
