% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmeans.R
\name{cmeans_elbowplot_sp}
\alias{cmeans_elbowplot_sp}
\title{cmeans_elbowplot_sp}
\usage{
cmeans_elbowplot_sp(object, name = "clean", variable)
}
\arguments{
\item{object}{an sp object \linkS4class{sp}.}

\item{name}{There are two options in total:
\itemize{ \item clean :
  Perform C-Means clustering using the slot clean_data.
  \item RID : Perform C-Means clustering using the slot RID_bootstrap.}}

\item{variable}{Group information of samples, which is a string corresponding
to one element in the column name of the sp object slot "sample_features".
The attribute of this column should be a character vector or a factor.}
}
\description{
Calculate the minimum centroid distances for a series of cluster
numbers and visualize the results to estimate the optimal number of clusters.
}
