% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmeans.R
\name{cmeans_heatmap_sp}
\alias{cmeans_heatmap_sp}
\title{cmeans_heatmap_sp}
\usage{
cmeans_heatmap_sp(
  object,
  name = "clean",
  min_mem = 0.5,
  cluster_name,
  col = NULL,
  show_row_names = F,
  show_column_names = F
)
}
\arguments{
\item{object}{an sp object \linkS4class{sp}.}

\item{name}{\itemize{ \item clean :
Visualize the C-Means clustering results from the slot clean_data.
\item RID : Visualize the C-Means clustering results from the slot RID_bootstrap.}}

\item{min_mem}{The membership values range between 0 and 1 and are used for feature selection.}

\item{cluster_name}{Group information of samples, which is a string corresponding
to one element in the column name of the sp object slot "sample_features".
The attribute of this column should be a character vector or a factor.}

\item{col}{Colors corresponding to value. A list contains named vectors.}

\item{show_row_names}{Whether show row names.}

\item{show_column_names}{Whether show column names.}
}
\value{
A Heatmap-class object.
}
\description{
Generate a heatmap to visualize the clustering outcomes of the C-Means clustering results.
}
