% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmeans.R
\name{cmeans_lineplot_sp}
\alias{cmeans_lineplot_sp}
\title{cmeans_lineplot_sp}
\usage{
cmeans_lineplot_sp(object, name = "clean", min_mem = 0.5, nrow = 1)
}
\arguments{
\item{object}{an sp object \linkS4class{sp}.}

\item{name}{\itemize{ \item clean :
Visualize the C-Means clustering results from the slot clean_data.
\item RID : Visualize the C-Means clustering results from the slot RID_bootstrap.}}

\item{min_mem}{The membership values range between 0 and 1 and are used for feature selection.}

\item{nrow}{Number of rows on the dividing surface.}
}
\value{
ggplot2 object.
}
\description{
Generate a heatmap to visualize the clustering outcomes of the C-Means clustering results.
}
