% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmeans.R
\name{cmeans_sp}
\alias{cmeans_sp}
\title{cmeans_sp}
\usage{
cmeans_sp(object, name = "clean", variable, ncluster = 10)
}
\arguments{
\item{object}{an sp object \linkS4class{sp}.}

\item{name}{There are two options in total:
\itemize{ \item clean :
  Perform C-Means clustering using the slot clean_data.
  \item RID : Perform C-Means clustering using the slot RID_bootstrap.}}

\item{variable}{Group information of samples, which is a string corresponding
to one element in the column name of the sp object slot "sample_features".
The attribute of this column should be a character vector or a factor.}

\item{ncluster}{number of clusters.}
}
\value{
Return the C-Means clustering results of features,
including a data frame feature_cluster containing category
information for different features, and an expression matrix of
different features in different variables.
}
\description{
Fuzzy C-Means clustering for proteins. Perform
clustering on different features based on the results of sample clustering.
}
