% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_jaccard_correlation.R
\name{correlation_sp}
\alias{correlation_sp}
\title{correlation_sp}
\usage{
correlation_sp(object, use = "pairwise.complete.obs", method = "pearson")
}
\arguments{
\item{object}{an sp object \linkS4class{sp}.}

\item{use}{\code{\link[stats]{cor}}}

\item{method}{\code{\link[stats]{cor}}}
}
\description{
Calculate the correlation index between a spot and its neighbors based on spatial
 adjacency relationships, and assign the results to a column in the sample_features slot.
}
