% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{creatsp}
\alias{creatsp}
\title{creatsp}
\usage{
creatsp(pathway, filetype = "xls")
}
\arguments{
\item{pathway}{The path where the original data is stored}

\item{filetype}{File extension}
}
\value{
a sp class
}
\description{
creat sp class
}
