% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolution.R
\name{deconvolution_sp}
\alias{deconvolution_sp}
\title{deconvolution_sp}
\usage{
deconvolution_sp(
  object,
  slot = "clean_data",
  sig_matrix,
  fromType = NULL,
  toType = NULL,
  organ = NULL
)
}
\arguments{
\item{object}{an sp object \linkS4class{sp}.}

\item{slot}{There are two options in total:
\itemize{ \item clean_data :
  Perform cell deconvolution using the slot clean_data.
  \item RID_bootstrap : Perform cell deconvolution using the slot RID_bootstrap.}}

\item{sig_matrix}{A matrix of expression profile of cells.}

\item{fromType}{The ID type of the row names in the selected expression matrix is default to NULL,
and it is used when converting the row names of the selected expression matrix to other ID type.}

\item{toType}{Corresponds to the fromType parameter, specifying the converted ID type.}

\item{organ}{Filter features with high correlation between transcriptome and proteome
in the mix_matrix based on the organ types of the samples in the matrix, which will be
used for deconvolution. If your sig_matrix is derived from RNA,
feature filtering is required prior to deconvolution.}
}
\value{
Assign a value to the slot deconvolution_result of the sp object.
}
\description{
Perform deconvolution for the sp object using the CIBERSORT method.
}
\details{
\itemize{ \item
The object and slot parameters are used to select the mix_matrix,
while the fromType and toType parameters are employed to convert
the IDs of the matrix for compatibility with the IDs of sig_matrix.
  \item When the organ parameter is not NULL, fromType cannot be NULL.}
}
