% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolution.R
\name{deconvolution_view}
\alias{deconvolution_view}
\title{deconvolution_view}
\usage{
deconvolution_view(object, nrow = 1, pal = NULL, legend_row = 1)
}
\arguments{
\item{object}{an sp object \linkS4class{sp}.}

\item{nrow}{Number of rows on the dividing surface.}

\item{pal}{The color palette to be used for coloring and filling by groups.}

\item{legend_row}{The number of rows in the legend. If there are too many types of groups,
you can appropriately increase the number of rows to improve the aesthetics of the plot.}
}
\value{
ggplot object.
}
\description{
Create a ggplot2-based pie chart for the deconvolution result.
}
