% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_plot.R
\name{density_plot_sp}
\alias{density_plot_sp}
\alias{density_plot_sp,sp-method}
\title{density_plot_sp}
\usage{
density_plot_sp(object)

\S4method{density_plot_sp}{sp}(object)
}
\arguments{
\item{object}{An \linkS4class{sp}.}
}
\value{
ggplot object.
}
\description{
Plot the probability density plot of sp object.
}
