% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dep.R
\name{dep_sp}
\alias{dep_sp}
\alias{dep_all_sp}
\title{dep_sp}
\usage{
dep_sp(
  object,
  slot = "clean_data",
  group,
  catagory1,
  catagory2,
  method = "wilcox"
)

dep_all_sp(object, slot = "clean_data", group, method = "wilcox")
}
\arguments{
\item{object}{an sp object \linkS4class{sp}.}

\item{slot}{There are two options in total:
\itemize{ \item clean_data :
  Perform differential calculation using the slot clean_data.
  \item RID_bootstrap : Perform differential calculation using the slot RID_bootstrap.}}

\item{group}{Group information of samples, which is a string corresponding
to one element in the column name of the sp object slot "sample_features".
The attribute of this column should be a character vector or a factor.}

\item{catagory1}{Corresponding to one or more groups in the group parameter,
this argument is used to select samples for comparison.}

\item{catagory2}{Corresponding to one or more groups in the group parameter,
this argument is used to select samples for comparison.}

\item{method}{There are three options in total:
\itemize{ \item wilcox :
  Differentially expressed proteins were identified using the wilcox test.
  \item t_test : Differentially expressed proteins were identified using the t_test test.
  \item moderated_t : Differentially expressed proteins were identified using the wilcox moderated t_test from the limma package..}}
}
\value{
A data frame containing statistical information on differentially expressed proteins.

A data frame containing statistical information on differentially expressed proteins.
}
\description{
Calculate differentially expressed proteins among different groups in spatial proteomics (sp) objects

Calculate differentially expressed proteins among all groups in a specific column
of sample_features within the spatial proteomics (sp) object.
}
