% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_plot_sp.R
\name{filter_plot_sp}
\alias{filter_plot_sp}
\title{filter_plot_sp}
\usage{
filter_plot_sp(object, times = 1, group = NULL, nrow = 1)
}
\arguments{
\item{object}{an sp object \linkS4class{sp}.}

\item{times}{The coefficient of the IQR value, with a default value of 1.}

\item{group}{Group information of samples, which is a string corresponding
to one element in the column name of the sp object slot "sample_features".
The attribute of this column should be a character vector or a factor.}

\item{nrow}{Number of rows on the dividing surface.}
}
\value{
ggplot object.
}
\description{
Draw violin plots of protein identification numbers and filter dashed lines.
}
