% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea_sp.R
\name{gsea_sp}
\alias{gsea_sp}
\title{gsea_sp}
\usage{
gsea_sp(
  object,
  protein_rank = NULL,
  dim_reduction_name,
  dim_value = c(1),
  pathway,
  fromType = NULL,
  toType = NULL,
  species = NULL
)
}
\arguments{
\item{object}{an sp object \linkS4class{sp}.}

\item{protein_rank}{Named vector of gene-level stats.}

\item{dim_reduction_name}{The dimensionality reduction objects stored in the slot dim_reductions of an sp-type object
can be either PCA-type data or PLS-DA-type data. such as "pca_clean".}

\item{dim_value}{A positive integer value, or a vector composed of positive integers.
When using the dimensionality reduction data contained within the sp object as the basis for estimating gsea enrichment scores,
the ranking must be based on the contribution of genes across different dimensions of the dimensionality reduction data. Therefore,
it is necessary to select dimension information: if PCA data is chosen, the value in the parameter must not exceed the number of samples;
if PLS-DA data is chosen, the value in the parameter must be either not exceed 2.}

\item{pathway}{a file path to a gmt format file, or a list storing vectors of different gene or protein IDs.}

\item{fromType}{The ID type of the row names in the selected expression matrix is default to NULL,
and it is used when converting the row names of the selected expression matrix to other ID type.}

\item{toType}{Corresponds to the fromType parameter, specifying the converted ID type.}

\item{species}{The name of the species-specific annotation package, such as "org.Hs.eg.db".}
}
\description{
Estimates gsea enrichment scores for the given sp object
}
\details{
If the protein_rank parameter is not NULL, the dimensionality reduction data contained
in the sp object will not be used, and the parameters dim_reduction_name and dim_value
will be ignored.
}
