% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_heatmap.R
\name{heatmap_sp}
\alias{heatmap_sp}
\title{heatmap_sp}
\usage{
heatmap_sp(
  object,
  name = "clean",
  dimn,
  cluster_name,
  col = NULL,
  show_row_names = F,
  show_column_names = F
)
}
\arguments{
\item{object}{an sp object \linkS4class{sp}.}

\item{name}{\itemize{ \item clean :
Visualize the pca results from the clean_data.
\item RID : Visualize the pca results of the data with individual differences removed through the bootstrap method (RID_bootstrap).
\item harmony : Visualize the pca results of the data with individual differences removed through the harmony method (corrective_PCA).}}

\item{dimn}{Select the expression matrix of the first n dimensions from the
pca results for visualization.}

\item{cluster_name}{Group information of samples, which is a string corresponding
to one element in the column name of the sp object slot "sample_features".
The attribute of this column should be a character vector or a factor.}

\item{col}{Colors corresponding to value. A list contains named vectors.}

\item{show_row_names}{Whether show row names.}

\item{show_column_names}{Whether show column names.}
}
\value{
A Heatmap-class object.
}
\description{
Generate a heatmap to visualize the clustering outcomes of the sp object.
}
