% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_jaccard_correlation.R
\name{jaccard_sp}
\alias{jaccard_sp}
\title{jaccard_sp}
\usage{
jaccard_sp(object)
}
\arguments{
\item{object}{an sp object \linkS4class{sp}.}
}
\description{
Calculate the Jaccard index between a spot and its neighbors based on spatial
 adjacency relationships, and assign the results to a column in the sample_features slot.
}
