% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jca.R
\name{jca_sp}
\alias{jca_sp}
\title{jca_sp}
\usage{
jca_sp(object, cluster, union = F)
}
\arguments{
\item{object}{an sp object \linkS4class{sp}.}

\item{cluster}{cluster information of samples, which is a string corresponding.
to one element in the column name "sample_features" of the sp object.}

\item{union}{default FALSE, if TRUE all individuals' neighbours list will be consolidated into one large list.}
}
\value{
if union=F, A matrix with class jcmulti with row and column names for observed and expected
counts, variance, and z-value. if union=T, Return a list containing the jcmulti matrix for each individual.
 expected counts, variance, and z-value.
}
\description{
Given the global counts in each cluster,
expected counts and variances are calculated under non-free sampling, and a z-value reported.
}
