% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_jca_visualization.R
\name{jca_visualization}
\alias{jca_visualization}
\title{jca_visualization}
\usage{
jca_visualization(x, colors = c("#2556A6", "#ffffff", "#EE2A29"), nrow = 1)
}
\arguments{
\item{x}{The result returned by the jca_sp function can be a matrix or a list.}

\item{colors}{A vector containing three colors.}

\item{nrow}{Number of rows on the dividing surface. If x is a matrix, then this parameter is useless.}
}
\description{
Visualize the output results of the jca_sp function.
}
