% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{mean_cv}
\alias{mean_cv}
\alias{mean_cv,sp-method}
\title{mean_cv}
\usage{
\S4method{mean_cv}{sp}(object)
}
\arguments{
\item{object}{A parameter object of one of the following classe sp or spatials}
}
\value{
a dataframe containing the mean and standard deviation of each sample
}
\description{
Calculate mean and cv
}
