% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moran.R
\name{moran_sp}
\alias{moran_sp}
\title{moran_sp}
\usage{
moran_sp(
  object,
  ssgsea = "no",
  scaled = T,
  na.rm = T,
  alternative = "two.sided"
)
}
\arguments{
\item{object}{an sp object \linkS4class{sp}.}

\item{ssgsea}{There are three options in total:
\itemize{ \item no :
  Default value, Calculate the Moran's I of features in the slot clean_data.
  \item clean : Calculate the Moran's I of the matrix named ssgsea_clean in ssgsea_list.
  \item RID : Calculate the Moran's I of the matrix named ssgsea_RID in ssgsea_list.}}

\item{scaled}{a logical indicating whether the Moran's I should be scaled so that it varies between -1 and +1  (default to TRUE).}

\item{na.rm}{a logical indicating whether missing values should be removed.}

\item{alternative}{a character string specifying the alternative hypothesis that is tested
against the null hypothesis of no phylogenetic correlation; must be of one "two.sided",
"less", or "greater", or any unambiguous abbrevation of these.}
}
\value{
A data frame contains various elements. See also \code{\link[ape]{Moran.I}}
}
\description{
Calculate the Moran's I of proteins or biological pathways
}
