% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvi.R
\name{mvi}
\alias{mvi}
\title{mvi}
\usage{
mvi(x, method = "impseqrob", nmr = 0)
}
\arguments{
\item{x}{A protein expression matrix (data.frame format), where the row names are protein
or gene IDs and the column names are sample names.}

\item{method}{Methods for missing value imputation: There are four methods in total,
namely impseqrob (Robust sequential imputation), impseq (Sequential imputation),
grr (Glmnet Ridge Regression), and seqknn (Sequential knn).
where the row names are protein or gene IDs and the column names are sample names.}

\item{nmr}{Non-missing rate. Using the non-missing rate as the threshold,
filter out proteins below this value before performing missing value imputation.}
}
\value{
A protein expression matrix with missing values imputed.
}
\description{
Impute missing values of protein expression matrix.
}
