% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvi.R
\name{mvi_sp}
\alias{mvi_sp}
\title{mvi_sp}
\usage{
mvi_sp(object, method = "impseqrob", nmr = 0)
}
\arguments{
\item{object}{an sp object \linkS4class{sp}.}

\item{method}{Methods for missing value imputation: There are four methods in total,
namely impseqrob (Robust sequential imputation), impseq (Sequential imputation),
grr (Glmnet Ridge Regression), and seqknn (Sequential knn).
where the row names are protein or gene IDs and the column names are sample names.}

\item{nmr}{Non-missing rate. Using the non-missing rate as the threshold,
filter out proteins below this value before performing missing value imputation.}
}
\value{
The sp object obtained after performing missing value imputation on clean_data.
}
\description{
Impute missing values for sp object slot clean_data.
}
\seealso{
\code{\link{mvi}}
}
