% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{normlization}
\alias{normlization}
\alias{normlization,sp-method}
\title{normlization}
\usage{
\S4method{normlization}{sp}(object, meancut, cvcut)
}
\arguments{
\item{object}{A parameter object of one of the following classe sp or spatials}

\item{meancut}{A parameter object of one of the following classe sp or spatials}

\item{cvcut}{A parameter object of one of the following classe sp or spatials}
}
\value{
A dataframe contains data that has been quality controlled and normalized
}
\description{
normalize raw data
}
