% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm.R
\name{normlization_sp}
\alias{normlization_sp}
\title{normlization_sp}
\usage{
normlization_sp(object)
}
\arguments{
\item{object}{an sp object \linkS4class{sp}.}
}
\description{
Normlize the slot clean_data of the sp object by Quantiles method.
}
