% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_point.R
\name{pca_point}
\alias{pca_point}
\title{pca_point}
\usage{
pca_point(x, group, pc = c(1, 2), pal = NULL, shape = "chull", legend_row = 1)
}
\arguments{
\item{x}{an prcomp object of class PCA.}

\item{group}{Group information of samples, which is a character vector or a factor.}

\item{pc}{A vector containing two numerical values, used to select two principal
components for display, with the first and second dimensions as the default.}

\item{pal}{The color palette to be used for coloring and filling by groups.}

\item{shape}{There are two options available: chull (convex hull) and ellipse.}

\item{legend_row}{The number of rows in the legend. If there are too many types of groups,
you can appropriately increase the number of rows to improve the aesthetics of the plot.}
}
\value{
ggplot object.
}
\description{
Create a ggplot2-based scatter plot of PCA outputs prcomp.
}
