% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_point.R
\name{pca_point_sp}
\alias{pca_point_sp}
\title{pca_point_sp}
\usage{
pca_point_sp(
  object,
  group,
  pc = c(1, 2),
  pal = NULL,
  shape = "chull",
  legend_row = 1,
  name = "clean"
)
}
\arguments{
\item{object}{an sp object \linkS4class{sp}.}

\item{group}{Group information of samples, which is a string corresponding
to one element in the column name of the sp object slot "sample_features".
The attribute of this column should be a character vector or a factor.}

\item{pc}{A vector containing two numerical values, used to select two principal
components for display, with the first and second dimensions as the default.}

\item{pal}{The color palette to be used for coloring and filling by groups.}

\item{shape}{There are two options available: chull (convex hull) and ellipse.}

\item{legend_row}{The number of rows in the legend. If there are too many types of groups,
you can appropriately increase the number of rows to improve the aesthetics of the plot.}

\item{name}{There are two options in total:
\itemize{ \item clean :
  Visualize the PCA results of the slot clean_data.
  \item RID : Visualize the PCA results of the slot RID_bootstrap.}}
}
\value{
ggplot object.
}
\description{
Create a ggplot2-based scatter plot of PCA for sp object.
}
\seealso{
\code{\link{pca_point}}
}
