% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plsda.R
\name{plsda_point_sp}
\alias{plsda_point_sp}
\title{plsda_point_sp
Create a ggplot2-based scatter plot of Plsda for sp object.}
\usage{
plsda_point_sp(object, name, col = NULL)
}
\arguments{
\item{object}{\linkS4class{sp}.}

\item{name}{There are two options in total:
\itemize{ \item clean :
  Visualize the plsda results of the original data.
  \item RID : Visualize the plsda results of the data with individual differences removed.}}

\item{col}{The color palette to be used for coloring and filling by groups.}
}
\value{
ggplot object.
}
\description{
plsda_point_sp
Create a ggplot2-based scatter plot of Plsda for sp object.
}
