% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection_rate.R
\name{ppd}
\alias{ppd}
\alias{ppd.data.frame}
\title{ppd (Probability of protein detection).}
\usage{
ppd(x, ...)

\method{ppd}{data.frame}(x, group, ...)
}
\arguments{
\item{x}{A protein expression matrix, where the row names are protein
or gene IDs and the column names are sample names.}

\item{...}{Arguments passed to other methods.}

\item{group}{information of samples, which is a character vector or factor.}
}
\value{
A data frame containing the detection rate of different proteins in each group.
}
\description{
Calculate the detection rate of proteins in different groups.
}
\examples{
\dontrun{
 group = rep(c("groupA", "groupA", "gruupB"), each = 15)

 set.seed(123)
 data <- data.frame(
 pro1 = sample(c(1:10, NA), 45, replace = TRUE),
 pro2 = sample(c(20:30, NA), 45, replace = TRUE),
 pro3 = sample(c(50:60, NA), 45, replace = TRUE)
 )

 data <- t(data) \%>\% as.data.frame()

 ppd(data,group)

}
}
