% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_harmony.R
\name{run_harmony_sp}
\alias{run_harmony_sp}
\alias{run_harmony_sp,sp-method}
\title{run_harmony_sp}
\usage{
run_harmony_sp(object, ...)

\S4method{run_harmony_sp}{sp}(object, dimn, group, theta = 2.5, sigma = 1.2)
}
\arguments{
\item{object}{An \linkS4class{sp}.}

\item{...}{Arguments passed to other methods.}

\item{dimn}{Select the expression matrix of the first n dimensions from the
principal component analysis result for harmony.}

\item{group}{Group information of samples, which is a string corresponding.
to one element in the column name "sample_features" of the sp object
The attribute of this column should be a character vector or a factor.}

\item{theta}{parameters for \code{\link[harmony]{RunHarmony.default}}}

\item{sigma}{parameters for \code{\link[harmony]{RunHarmony.default}}}
}
\description{
Performs the Harmony algorithm directly on the given sp object.
}
