% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_pca.R
\name{run_pca}
\alias{run_pca}
\title{run_pca}
\usage{
run_pca(x, filter_by = "sd", rank = 0.5)
}
\arguments{
\item{x}{A protein expression matrix (data.frame format), where the row names are protein
or gene IDs and the column names are sample names.}

\item{filter_by}{Select features with large fluctuations for principal components analysis,
The standard deviation (sd) or coefficient of variation (cv) can be used for feature selection.}

\item{rank}{A numeric value between 0 and 1. Select features with the largest fluctuations based on percentage.
By default, select the top 50\% of features for principal components analysis.}
}
\value{
returns a list with class "prcomp".
}
\description{
Performs a principal components analysis on the given protein matrix and
returns the results as an object of class prcomp.
}
\details{
If your protein expression matrix is in logarithmic form (e.g., after log2 transformation),
we recommend using the standard deviation as the screening criterion.
This is because the difference between data represents a relative change in fold (multiples),
rather than the absolute value of abundance.
}
