% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plsda.R
\name{run_plsda_sp}
\alias{run_plsda_sp}
\title{run_plsda_sp}
\usage{
run_plsda_sp(object, filter_by = "sd", rank = 0.5, group, name = "clean")
}
\arguments{
\item{object}{an sp object \linkS4class{sp}.}

\item{filter_by}{Select features with large fluctuations for principal components analysis,
The standard deviation (sd) or coefficient of variation (cv) can be used for feature selection.}

\item{rank}{Select features with the largest fluctuations based on percentage.
By default, select the top 50\% of features for principal components analysis.}

\item{group}{Group information of samples, which is a string corresponding
to one element in the column name of the sp object slot "sample_features".
The attribute of this column should be a character vector or a factor.}

\item{name}{There are two options in total:
\itemize{ \item clean :
  Perform plsda dimensionality reduction using the slot clean_data, and store the returned result in dim_reductions[["plsda_clean"]].
  \item RID : Perform plsda dimensionality reduction using the slot RID_bootstrap, and store the returned result in dim_reductions[["pca_RID"]].}}
}
\description{
Performs plsda analysis on the sp object.
}
\seealso{
\code{\link{run_plsda}}
}
