% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_umap.R
\name{run_umap}
\alias{run_umap}
\title{run_umap}
\usage{
run_umap(x, dimn)
}
\arguments{
\item{x}{an prcomp object of class PCA.}

\item{dimn}{Select the expression matrix of the first n dimensions from the
pca result for UMAP dimensional reduction.}
}
\value{
return object of class umap.
}
\description{
Perform the Uniform Manifold Approximation and Projection (UMAP) dimensional reduction technique.
}
