% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_umap.R
\name{run_umap_sp}
\alias{run_umap_sp}
\title{run_umap_sp}
\usage{
run_umap_sp(object, dimn, name = "clean")
}
\arguments{
\item{object}{an sp object \linkS4class{sp}.}

\item{dimn}{Select the expression matrix of the first n dimensions from the
pca result for UMAP dimensional reduction.}

\item{name}{There are three options in total:
\itemize{ \item clean :
  The PCA results of the original data (the slot clean_data).
  \item RID : The PCA results of the data with individual differences removed through the bootstrap method.
   \item harmony : The PCA results of the data with individual differences removed through the harmony method.}}
}
\description{
Performs Uniform Manifold Approximation and Projection (UMAP) dimensional reduction on the sp object.
}
\details{
UMAP dimensionality reduction is based on PCA results.
}
\seealso{
\code{\link{run_umap}}
}
