% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\docType{class}
\name{sp-class}
\alias{sp-class}
\alias{sp}
\title{the sp class}
\description{
class designed for one experiment
}
\section{Slots}{

\describe{
\item{\code{rawdata}}{data.frame.}

\item{\code{features}}{data.frame.}

\item{\code{scaledata}}{data.frame.}

\item{\code{dim}}{list.}
}}

