% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_jca_visualization.R
\name{sp_visualization}
\alias{sp_visualization}
\title{sp_visualization}
\usage{
sp_visualization(
  object,
  variale,
  nrow = 1,
  ssgsea = "no",
  continuous_variable = T,
  fills = c("#2556A6", "#ffffff", "#EE2A29")
)
}
\arguments{
\item{object}{an sp object \linkS4class{sp}.}

\item{variale}{The name of the variable to be displayed can be the name of a column
in the sample_features data frame, the name of a row in the clean_data data frame,
or the name of a pathway in the ssgsea results.}

\item{nrow}{Number of rows on the dividing surface.}

\item{ssgsea}{There are three options in total:
\itemize{ \item no :
  Default value, indicating that the variable to be visualized is not part of the ssgsea results.
  \item clean : The variable to be visualized is derived from the matrix named ssgsea_clean in ssgsea_list.
  \item RID : The variable to be visualized is derived from the matrix named ssgsea_RID in ssgsea_list.}}

\item{continuous_variable}{A logical value, defaulting to TRUE, which determines whether the variable
is continuous or discrete.}

\item{fills}{If the continuous_variable parameter is TRUE, input a vector containing three colors;
if continuous_variable is FALSE, input a number of colors corresponding to the number of categories
of the variable to be visualized.}
}
\value{
ggplot object.
}
\description{
Visualize the global spatial proteome.
}
