% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\docType{class}
\name{spatials-class}
\alias{spatials-class}
\alias{spatials}
\title{The spatials class}
\description{
An S4 object that stores multiple sp objects, but only one sp object is activated at the same time
}
\section{Slots}{

\describe{
\item{\code{experiment}}{list.}

\item{\code{active}}{character.}
}}

