% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssgsea_sp.R
\name{ssgsea_sp}
\alias{ssgsea_sp}
\title{ssgsea_sp}
\usage{
ssgsea_sp(
  object,
  name = "clean",
  pathway,
  fromType = NULL,
  toType = NULL,
  species = NULL
)
}
\arguments{
\item{object}{an sp object \linkS4class{sp}.}

\item{name}{There are two options in total:
\itemize{ \item clean :
  Perform ssgsea enrichment analysis using the slot clean_data, and store the returned result in ssgsea_list[["ssgsea_list_clean"]].
  \item RID : Perform ssgsea enrichment analysis using the slot RID_bootstrap, and store the returned result in ssgsea_list[["ssgsea_list_RID"]].}}

\item{pathway}{a file path to a gmt format file, or a list storing vectors of different gene or protein IDs.}

\item{fromType}{The ID type of the row names in the selected expression matrix is default to NULL,
and it is used when converting the row names of the selected expression matrix to other ID type.}

\item{toType}{Corresponds to the fromType parameter, specifying the converted ID type.}

\item{species}{The name of the species-specific annotation package, such as "org.Hs.eg.db".}
}
\description{
Estimates ssgsea enrichment scores for the given sp object
}
\details{
The ID type of the selected expression matrix should be consistent with those contained in the input pathway.
}
