% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umap_point.R
\name{umap_point}
\alias{umap_point}
\alias{umap_point.umap}
\title{umap_point}
\usage{
umap_point(x, ...)

\method{umap_point}{umap}(x, group, pal = NULL, shape = "chull", legend_row = 1, ...)
}
\arguments{
\item{x}{object of class umap.}

\item{...}{Arguments passed to other methods.}

\item{group}{Group information of samples, which is a character vector or factor.}

\item{pal}{The color palette to be used for coloring and filling by groups.}

\item{shape}{There are two options available: chull (convex hull) and NULL.}

\item{legend_row}{The number of rows in the legend. If there are too many types of groups,
you can appropriately increase the number of rows to improve the aesthetics of the plot.}
}
\value{
ggplot object.
}
\description{
Create a ggplot2-based scatter plot for Uniform Manifold Approximation and Projection (UMAP) dimensional reduction.
}
