% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umap_point.R
\name{umap_point_sp}
\alias{umap_point_sp}
\title{umap_point_sp}
\usage{
umap_point_sp(
  object,
  group,
  pal = NULL,
  shape = NULL,
  legend_row = 1,
  name = "clean"
)
}
\arguments{
\item{object}{an sp object \linkS4class{sp}.}

\item{group}{Group information of samples, which is a string corresponding
to one element in the column name of the sp object slot "sample_features".
The attribute of this column should be a character vector or a factor.}

\item{pal}{The color palette to be used for coloring and filling by groups.}

\item{shape}{There are two options available: chull (convex hull) and ellipse.}

\item{legend_row}{The number of rows in the legend. If there are too many types of groups,
you can appropriately increase the number of rows to improve the aesthetics of the plot.}

\item{name}{There are three options in total:
\itemize{ \item clean :
  TVisualize the umap results of the original data. (the umap_clean).
  \item RID : Visualize the umap results of the data with individual differences removed through the bootstrap method. (umap_RID)
   \item harmony : Visualize the umap results of the data with individual differences removed through the harmony method. (umap_harmony)}}
}
\value{
ggplot object.
}
\description{
Create a ggplot2-based scatter plot of umap for sp object.
}
\seealso{
\code{\link{umap_point}}
}
