% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_plot.R
\name{density_plot}
\alias{density_plot}
\title{density_plot}
\usage{
density_plot(x)
}
\arguments{
\item{x}{A protein expression matrix (data.frame format), where the row names are protein
or gene IDs and the column names are sample names.}
}
\value{
ggplot object.
}
\description{
Plot the probability density plot of the protein expression matrix.
}
