% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_plot.R
\name{density_plot_sp}
\alias{density_plot_sp}
\title{density_plot_sp}
\usage{
density_plot_sp(object)
}
\arguments{
\item{object}{an sp object \linkS4class{sp}.}
}
\value{
ggplot object.
}
\description{
Plot the probability density plot of sp object.
}
