% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_sample.R
\name{filter_sample}
\alias{filter_sample}
\title{filter_sample}
\usage{
filter_sample(x, times = 1, group = NULL)
}
\arguments{
\item{x}{A protein expression matrix (data.frame format), where the row names are protein
or gene IDs and the column names are sample names.}

\item{times}{The coefficient of the IQR value, with a default value of 1.}

\item{group}{Group information of samples, which is a character vector or factor.}
}
\value{
A data.frame with Some Columns (Samples) Filtered Out.
}
\description{
Filter out samples where the number of identified proteins
is less than (25th percentile minus n times IQR).
}
