% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_var_bar.R
\name{pca_bar_sp}
\alias{pca_bar_sp}
\alias{pca_bar_sp,sp-method}
\title{pca_bar_sp}
\usage{
pca_bar_sp(object, ...)

\S4method{pca_bar_sp}{sp}(
  object,
  addlabels = TRUE,
  ncp = 15,
  ggtheme = theme_classic(),
  name = "RAW"
)
}
\arguments{
\item{object}{An \linkS4class{sp}.}

\item{...}{Arguments passed to other methods.}

\item{addlabels}{logical value. If TRUE, labels are added at the top of bars or
points showing the information retained by each dimension.}

\item{ncp}{a numeric value specifying the number of dimensions to be shown.
The attribute of this column should be a character vector or a factor.}

\item{ggtheme}{function, ggplot2 theme name.}

\item{name}{There are two options in total:
\itemize{ \item RAW :
  Visualize the PCA results of the original data.
  \item RID : Visualize the PCA results of the data with individual differences removed.}}
}
\description{
Visualize the variances of pca dimensions for the given sp object.
}
