% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_var.R
\name{pca_var_sp}
\alias{pca_var_sp}
\title{pca_var_sp}
\usage{
pca_var_sp(object, pc = c(1, 2), n = 10, name = "clean")
}
\arguments{
\item{object}{an sp object \linkS4class{sp}.}

\item{pc}{A vector containing two numerical values, used to select two principal
components for display, with the first and second dimensions as the default.}

\item{n}{The number of features to display, including n positively correlated}

\item{name}{There are two options in total:
\itemize{ \item clean :
  Visualize the PCA results of the slot clean_data.
  \item RID : Visualize the PCA results of the slot RID_bootstrap.}}
}
\value{
ggplot object.
}
\description{
Visualize the correlation between features and different principal components using ggplot2 for the sp object.
}
\seealso{
\code{\link{pca_var}}
}
