% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection_rate.R
\name{ppd_sp}
\alias{ppd_sp}
\title{ppd_sp}
\usage{
ppd_sp(object, group)
}
\arguments{
\item{object}{an sp object \linkS4class{sp}.}

\item{group}{Group information of samples, which is a string corresponding
to one element in the column name of the sp object slot "sample_features".
The attribute of this column should be a character vector or a factor.}
}
\value{
A data frame containing the detection rate of different proteins in each group.
}
\description{
Calculate the detection rate of proteins in different groups for sp object.
}
\seealso{
\code{\link{ppd}}
}
