% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_umap.R
\name{run_umap_sp}
\alias{run_umap_sp}
\alias{run_umap_sp,sp-method}
\title{run_umap_sp}
\usage{
run_umap_sp(object, ...)

\S4method{run_umap_sp}{sp}(object, dimn, name = "RAW")
}
\arguments{
\item{object}{An \linkS4class{sp}.}

\item{...}{Arguments passed to other methods.}

\item{dimn}{Select the expression matrix of the first n dimensions from the
principal component analysis for UMAP dimensional reduction.}

\item{name}{There are two options in total:
\itemize{ \item RAW :
  Visualize the PCA results of the original data.
  \item RID : Visualize the PCA results of the data with individual differences removed.}}
}
\description{
Performs Uniform Manifold Approximation and Projection (UMAP) dimensional reduction on the given sp object.
}
\seealso{
\code{\link{run_umap.prcomp}}
}
