% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\docType{class}
\name{sp-class}
\alias{sp-class}
\alias{sp}
\title{the sp class}
\description{
class designed for one experiment.
}
\section{Slots}{

\describe{
\item{\code{rawdata}}{data.frame.}

\item{\code{sample_features}}{data.frame.}

\item{\code{protein_features}}{data.frame.}

\item{\code{clean_data}}{data.frame.}

\item{\code{dim_reductions}}{list.}

\item{\code{corrective_PCA}}{data.frame.}

\item{\code{RID_bootstrap}}{data.frame.}

\item{\code{cmeans_list}}{"list".}

\item{\code{ssgsea_list}}{list.}

\item{\code{GSEA_list}}{list.}

\item{\code{adj_matrix_list}}{list.}

\item{\code{list_neighbour}}{list.}

\item{\code{combine_listw}}{ANY.}

\item{\code{deconvolution_result}}{data.frame.}
}}

