% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\docType{class}
\name{sp-class}
\alias{sp-class}
\alias{sp}
\title{the sp class}
\description{
class designed for one experiment.
}
\section{Slots}{

\describe{
\item{\code{rawdata}}{data.frame.}

\item{\code{sample_features}}{data.frame.}

\item{\code{protein_features}}{data.frame.}

\item{\code{cleaned_data}}{data.frame.}

\item{\code{dim_reductions}}{list.}

\item{\code{corrective_PCA}}{data.frame.}

\item{\code{RID_bootstrap}}{data.frame.}
}}

