% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umap_point.R
\name{umap_point_sp}
\alias{umap_point_sp}
\alias{umap_point_sp,sp-method}
\title{umap_point_sp}
\usage{
umap_point_sp(object, ...)

\S4method{umap_point_sp}{sp}(
  object,
  group,
  pal = NULL,
  shape = "chull",
  legend_row = 1,
  name = "RAW"
)
}
\arguments{
\item{object}{An \linkS4class{sp}.}

\item{...}{Arguments passed to other methods.}

\item{group}{Group information of samples, which is a string corresponding.
to one element in the column name "sample_features" of the sp object
The attribute of this column should be a character vector or a factor.}

\item{pal}{The color palette to be used for coloring and filling by groups.}

\item{shape}{There are two options available: chull (convex hull) and ellipse.}

\item{legend_row}{The number of rows in the legend. If there are too many types of groups,
you can appropriately increase the number of rows to improve the aesthetics of the plot.}

\item{name}{There are two options in total:
\itemize{ \item RAW :
  Visualize the umap results of the original data.
  \item RID : Visualize the umap results of the data with individual differences removed.}}
}
\value{
ggplot object.
}
\description{
Create a ggplot2-based scatter plot of umap for sp object.
}
\seealso{
\code{\link{umap_point.umap}}
}
